Lesson 2 (Challenge 1 - Excel Formula)
On this challenge, it will require us to use basic formulas in Excel to accomplish the following tasks: 

Click the Challenge tab in the bottom-left of the workbook.
Create a formula in cell D4 that multiplies the quantity in B4 by the price per unit in cell C4.
Use the fill handle to copy the formula in cell D4 to cells D5:D7.
Change the price per unit for the fried plantains in cell C6 to $2.25. Notice that the line total automatically changes as well.
Edit the formula for the total in cell D8 so it also adds cell D7.