Lesson 2 (Challenge 2 - Excel Formula)
In cell D7, create a formula that calculates the tax for the invoice. Use a sales tax rate of 7.5%.
In cell D8, create a formula that finds the total for the order. In other words, this formula should add cells D3:D7.
In cell D9 create a formula that calculates the total after a 10% discount.